Using url parameters to manage page and search
Through the use of the client side hooks
useSearchParams() and the searchParams prop of
the Next.js page component, I was able to get
info from the URL that could be used to alter
UI and display queried data.

I also used debouncing to prevent requests to
be sent to the server per each keystroke in the
search, which helps to optimize the web app.