Fixed 'Module Not Found' issue within middleware
For some reason, the Next.js service was unable to
locate the NextAuth module when trying to import
it via 'next-auth' from the project's root.

In order to make it work, I simply gave the
import code a relative path, after which pnpm was
able to compile and run the code quickly.